requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.2
daphne==4.0.0
Django==5.0.1
django-cors-headers==4.3.1
django-eventstream==5.1.0
django-grip==3.5.0
djangorestframework==3.14.0
gripcontrol==4.2.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.0
pillow==10.2.0
pubcontrol==3.5.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.1
zope.interface==6.1
